+++ src/stream/upload_stdin.rs
@@ -2,9 +2,9 @@ use crate::{
     cli::globals::GlobalArgs,
     s3::S3,
     stream::{
-        InitialStreamParams, STDIN_BUFFER_SIZE, Stream, complete_multipart_upload, compress_chunk,
+        InitialStreamParams, STDIN_BUFFER_SIZE, Stream, complete_multipart_upload,
         create_initial_stream, get_key, initiate_multipart_upload, maybe_upload_part,
-        setup_progress, upload_final_part, write_to_stream,
+        setup_stream_progress, upload_final_part, write_to_stream,
     },
 };
 use anyhow::{Result, anyhow};
@@ -25,19 +25,14 @@ pub async fn stream_stdin(
     tmp_dir: PathBuf,
     globals: GlobalArgs,
 ) -> Result<String> {
-    // use .zst extension if compress option is set
     let key = get_key(object_key, globals.compress, globals.encrypt);
 
-    let mut meta = meta.unwrap_or_default();
-
-    if globals.compress {
-        meta.insert("Content-Type".to_string(), "application/zstd".to_string());
-    }
+    let meta = meta.unwrap_or_default();
 
     // S3 setup
     let upload_id = initiate_multipart_upload(s3, &key, acl, meta).await?;
 
-    let progress_sender = setup_progress(quiet, None).await;
+    let progress_sender = setup_stream_progress(quiet).await;
 
     // Create initial stream
     let first_stream: Stream = create_initial_stream(InitialStreamParams {
@@ -55,18 +50,9 @@ pub async fn stream_stdin(
         .try_fold(
             first_stream,
             |mut current_upload_state_acc, chunk| async move {
-                if globals.compress {
-                    // If compression is enabled, compress the current chunk
-                    let data = compress_chunk(chunk).await?;
-
-                    // Write the compressed chunk to our internal buffer/temp file
-                    write_to_stream(&mut current_upload_state_acc, &data)
-                        .map_err(|e| anyhow!("Error writing compressed chunk to stream: {e}"))?;
-                } else {
-                    // If compression is not enabled, write the raw chunk
-                    write_to_stream(&mut current_upload_state_acc, &chunk)
-                        .map_err(|e| anyhow!("Error writing raw chunk to stream: {e}"))?;
-                }
+                // write the raw chunk
+                write_to_stream(&mut current_upload_state_acc, &chunk)
+                    .map_err(|e| anyhow!("Error writing raw chunk to stream: {e}"))?;
 
                 // Check if a part needs to be uploaded to S3
                 maybe_upload_part(&mut current_upload_state_acc, STDIN_BUFFER_SIZE).await?;