+++ src/stream/upload_stdin_compressed.rs
@@ -1,32 +1,33 @@
-use crate::stream::{
-    FileStreamUpload, InitialStreamParams, STDIN_BUFFER_SIZE, Stream, complete_multipart_upload,
-    compress_chunk, create_initial_stream, get_key, initiate_multipart_upload, maybe_upload_part,
-    setup_stream_progress, upload_final_part, write_to_stream,
+use crate::{
+    cli::globals::GlobalArgs,
+    s3::S3,
+    stream::{
+        InitialStreamParams, STDIN_BUFFER_SIZE, Stream, complete_multipart_upload, compress_chunk,
+        create_initial_stream, get_key, initiate_multipart_upload, maybe_upload_part,
+        setup_stream_progress, upload_final_part, write_to_stream,
+    },
 };
 use anyhow::{Result, anyhow};
 use futures::stream::TryStreamExt;
-use tokio::fs::File;
+use std::{collections::BTreeMap, path::PathBuf};
+use tokio::io::stdin;
 use tokio_util::codec::{BytesCodec, FramedRead};
 
-/// Read file in chunks of 512MB
+/// Read from STDIN, compress the data using zstd, and upload in chunks
 /// # Errors
 /// Will return an error if the upload fails
-pub async fn stream_compressed(request: FileStreamUpload<'_>) -> Result<String> {
-    let FileStreamUpload {
-        s3,
-        object_key,
-        acl,
-        meta,
-        quiet,
-        tmp_dir,
-        globals,
-        file_path,
-    } = request;
-
-    // use .zst extension if compress option is set
+pub async fn stream_stdin_compressed(
+    s3: &S3,
+    object_key: &str,
+    acl: Option<String>,
+    meta: Option<BTreeMap<String, String>>,
+    quiet: bool,
+    tmp_dir: PathBuf,
+    globals: GlobalArgs,
+) -> Result<String> {
+    // use .zst extension
     let key = get_key(object_key, globals.compress, globals.encrypt);
 
-    // Add Content-Type application/zstd
     let mut meta = meta.unwrap_or_default();
     meta.insert("Content-Type".to_string(), "application/zstd".to_string());
 
@@ -46,21 +47,17 @@ pub async fn stream_compressed(request: FileStreamUpload<'_>) -> Result<String> 
         header_data: None,
     })?;
 
-    let file = File::open(file_path).await?;
-
-    // The accumulator for try_fold is a tuple: (UploadStream, EncryptorBE32).
-    // After the fold, we map the Result to extract only the UploadStream state.
-    let mut stream = FramedRead::new(file, BytesCodec::new())
-        .map_err(|e| anyhow!("Error reading file chunk: {e}"))
+    let mut stream = FramedRead::new(stdin(), BytesCodec::new())
+        .map_err(|e| anyhow!("Error reading STDIN chunk: {e}"))
         .try_fold(
             first_stream,
             |mut current_upload_state_acc, chunk| async move {
                 // Compress the current chunk
-                let compress_data = compress_chunk(chunk).await?;
+                let data = compress_chunk(chunk).await?;
 
-                // Write the encrypted chunk to our internal buffer/temp file
-                write_to_stream(&mut current_upload_state_acc, &compress_data)
-                    .map_err(|e| anyhow!("Error writing chunk to stream: {e}"))?;
+                // Write the compressed chunk to our internal buffer/temp file
+                write_to_stream(&mut current_upload_state_acc, &data)
+                    .map_err(|e| anyhow!("Error writing compressed chunk to stream: {e}"))?;
 
                 // Check if a part needs to be uploaded to S3
                 maybe_upload_part(&mut current_upload_state_acc, STDIN_BUFFER_SIZE).await?;