+++ src/stream/upload_stdin_compressed_encrypted.rs
@@ -4,7 +4,7 @@ use crate::{
     stream::{
         InitialStreamParams, STDIN_BUFFER_SIZE, Stream, complete_multipart_upload, compress_chunk,
         create_initial_stream, create_nonce_header, encrypt_chunk, get_key, init_encryption,
-        initiate_multipart_upload, maybe_upload_part, setup_progress, upload_final_part,
+        initiate_multipart_upload, maybe_upload_part, setup_stream_progress, upload_final_part,
         write_to_stream,
     },
 };
@@ -46,7 +46,7 @@ pub async fn stream_stdin_compressed_encrypted(
     // S3 setup
     let upload_id = initiate_multipart_upload(s3, &key, acl, meta).await?;
 
-    let progress_sender = setup_progress(quiet, None).await;
+    let progress_sender = setup_stream_progress(quiet).await;
 
     // Initialize encryption
     let (cipher, nonce_bytes) = init_encryption(encryption_key);