+++ src/stream/upload_stdin_encrypted.rs
@@ -4,7 +4,7 @@ use crate::{
     stream::{
         InitialStreamParams, STDIN_BUFFER_SIZE, Stream, complete_multipart_upload,
         create_initial_stream, create_nonce_header, encrypt_chunk, get_key, init_encryption,
-        initiate_multipart_upload, maybe_upload_part, setup_progress, upload_final_part,
+        initiate_multipart_upload, maybe_upload_part, setup_stream_progress, upload_final_part,
         write_to_stream,
     },
 };
@@ -42,7 +42,7 @@ pub async fn stream_stdin_encrypted(
     );
 
     let upload_id = initiate_multipart_upload(s3, &key, acl, meta).await?;
-    let progress_sender = setup_progress(quiet, None).await;
+    let progress_sender = setup_stream_progress(quiet).await;
 
     let (cipher, nonce_bytes) = init_encryption(encryption_key);
     let encryptor = EncryptorBE32::from_aead(cipher, (&nonce_bytes).into());